app/requirements.txt
python-pptx>=0.6.23
jinja2>=3.1.4
pyyaml>=6.0.2
regex>=2024.5.15
rich>=13.7.1
Pillow>=10.2.0
streamlit>=1.37.0

